#25184 factory more test
